Added getUserByEmail to src helpers from exporting tasks helpers, added button form submission in friends handlebars, added css to fix form formatting
